Refactoring tensorflow_cnn.chunks/ensemble_cnns.chunks into tools.chunks, also moving unit tests.
